cmd: ufetch: avoid truncating storage sizes on 32-bit systems
Storage capacities above 4 GiB are truncated when stored in size_t
on 32-bit systems. Use a 64-bit accumulator and promote the block
count before multiplication so 32-bit LBA configurations are also
handled correctly.

Fixes: 86d462c05d57 ("cmd: add a fetch utility")

+++ cmd/ufetch.c
@@ -95,7 +95,7 @@ static int do_ufetch(struct cmd_tbl *cmdtp, int flag, int argc,
 	const char *model, *compatible;
 	char *ipaddr;
 	int n_cmds, n_cpus = 0, compatlen;
-	size_t size = 0;
+	u64 size = 0;
 	ofnode np;
 	bool skip_ascii = false;
 
@@ -221,7 +221,7 @@ static int do_ufetch(struct cmd_tbl *cmdtp, int flag, int argc,
 			ret = uclass_find_device_by_seq(UCLASS_BLK, line - STORAGE, &dev);
 			if (!ret && dev) {
 				desc = dev_get_uclass_plat(dev);
-				size = desc->lba * desc->blksz;
+				size = (u64)desc->lba * desc->blksz;
 				printf("%4s %d: " RESET, blk_get_uclass_name(desc->uclass_id),
 					desc->lun);
 				if (size)
